Make vtkCompleteArrays handle vtkDataObjectTrees
Prior to this change, vtkCompleteArrays operated only on singular
datasets. When it was to be run on composite data, it relied on the
composite data pipeline to execute on each block. Blocks without data
are skipped by the pipeline, but in a parallel setting, this is a
problem because blocks may be null on one processor but not null on
another processor. Because vtkCompleteArrays uses an AllReduce(), this
could lead to an MPI communication mismatch and hence a crash.

Changing this class to handle vtkDataObjectTrees directly solves this
problem by making sure each block is processed regardless of whether
it is null or not. Hence all calls to AllReduce() are synchronized
across processors, avoiding the crash.