Fix build issues for dependent projects.
When client-server wrapping was generated for multiple libraries,
each of which contained a source file of the same name (perhaps
in a different namespace or include directory), the generated
wrapper files would have the same name and path causing CMake
to error out (since there were multiple rules to build the same
file). Add the library name being wrapped to the path holding the
generated wrapper file to prevent this.