Adding a Python wrapped method to export a transfer function
Note that the exported transfer function is in the JSON format and that it
does not affect the preset TFs. The idea is that if we're generating a
lot of TFs in an automated fashion we want to be able to save them
but we don't want to have them go into our stash of preset TFs
because some of them may be good and some may be worthless.