Enable specification of representation type to create
Provide a signature of Show() member function that takes a
representation type name that specifies which representation to
instantiate for a source proxy. If the type name is set to nullptr,
instantiate the representation the way it was previously.

Refactor code so that all the logic for creating a representation
is in the new Show() member function.