Specify representation type when Show() is traced
Make sure the same representation proxy is created when a script is
run as the one created when the trace was recorded. In certain cases,
such as a filter that produces a multiblock dataset, the default
representation depends on the types of datasets in the leaf nodes. If
a traced Python script is run and one of the filters produces a
multiblock dataset with empty leaf nodes, it is possible for a
different representation to be created for it than when the trace was
recorded.

For example, the Extract Selection filter may produce a multiblock
dataset of unstructured grids when a trace is recorded, and this gets
represented by an UnstructuredGridRepresentation. When the script is
later run, it is possible the selection comes up empty, in which case
the multiblock dataset leaves will be empty and a
GeometryRepresentation will be created instead. When this occurs,
properties traced from the UnstructuredGridRepresentation may be set
on the GeometryRepresentation, but since the GeometryRepresentation
doesn't have some of the properties the UnstructuredGridRepresentation
has, the traced script will terminate with errors.

Fixes #19491.