Do not override previously set closeEvent state
ParaView allows plugins to connect to the main window's close() event, allowing
them to reject the close event if necessary. This MR fixes a bug where ParaView
would then override the decision made by listeners to this event. With this
update, ParaView will simply pass the state of the closeEvent through (ParaView
still closes when prompted, but it does not override requests to cancel the
close event).