Spreadsheet: handle field data with variable lengths
Fixes issues with spreadsheet view when showing field data with arrays
having different lengths. When dealing with field data,
vtkAttributeDataToTableFilter now adds a new array called
`<arrayname>___vtkValidMask__` for each array that is padded.
vtkSpreadSheetView/pqSpreadSheetViewModel use this new array to
determine if a cell in the table is empty or has valid data.

Fixes #19588. Also fixes #17812 using a more reliable approach that
the fix in 9cc67986.