Fix array association of pqArraySelectorPropertyWidget
It was set to the domain association in the vtkSMArrayListDomain, but
this corresponds to the original array association. However, that
original array association is not always the one that is required by
the source. E.g., the Contour filter requires "point" data arrays, and
if the input contains cell data arrays and the AutoConvertProperties
setting is enabled, cell data arrays will appear in the widget. Cell
data arrays are not allowed by the Contour filter, and with the prior
setup, cell data arrays would be passed to it. Instead, we need to
pass the point data array that is autoconverted from the cell data
array.

To remedy this, change the array association to the field association
from the vtkSMArrayListDomain. This ensures the required array type is
set where applicable.