cpstate: cleanup globals
CoProcessingFullWorkflow tests at times failed to export state
correctly. The freqs in the generated script were empty at times causing
the test to fail since no images were output when the waveletdriver was
run. While it's unclear why, I suspect it's due to the way the
global state was managed. Cleaned it up to see if that helps.

`pqTemporalExportReaction` for some odd reason uses the `cpstate`. That
should be cleaned up. For now updating the code to init and cleanup
globals after pqTemporalExportReaction is done with the `cpstate`.