add temporal processing to Catalyst and add an ex post facto demo
This change inserts an optional vtkTemporalCacheFilter
between the simulation adaptor produced vtkDataObjects and the
processing pipeline. The cache then acts as a time varying source
such that Catalyst pipelines can apply temporal filters.

The API for the feature is demonstrated by the new CatalystExample.
This includes use in a C++ pipeline, and two Python Pipelines.
The second Python Pipeline is unique in that it demonstrates an
after the fact or "ex post facto" trigger. See comments in FEDriver
for how to run the example.