spack: cleanup local version code
We no longer use git-describe to determine version. That was not always
realiable e.g. the source was not a git checkout or when the checkout
was missing updated tags (which happened when checkout out forks).
Instead, now we use the version.txt to determine the version number and
suffix it with `-local`.