Matching new htg API with cell data
Hypertree grids used to have point data, which brought
confusion on how to deal with it in ParaView, because it
is cell data in the primal grid, and point data in the dual.
As of today, the data stored in hypertree grids is
represented as cell data.

This commit also fixes a few XML issues regarding htg filters
which were broken by this new API.