Use Breadth First Traversal algorithm for SIL selection
For big cases with a lot of nodes, the SIL does not behave well. If it builds a tree with lots of cross links. When the selection algorithm that relies on depth first traversal is iterating, it will slip through the links unsynchronizedly leading to multiple visit to same nodes. With the proposed breadth first traversal strategy, when the selection reach a link, all the links at the same depth level will be traversed immediately afterwards, leading to coherent updates in the neighbor tree branches and saving us from multiple visit to the same nodes.

CGNS files with lot of nodes and families are concerned (see #19333)