Prevent accidental display of color map editor histogram
When the "Automatically recompute data for histogram" option is on but
the "Display data histogram" option is off, the histogram would be
reshown if the color map range was changed. Fix this so that the
histogram remains hidden when the display option is turned off,
regardless of the state of the automatic update setting.