Implementation of resample to htg with mpi
This is the first step towards a multi-process
aware resample to hyper tree grid filter.
This filter works, however, if the extrapolate data
flag is on, the result with MPI will not be numerically
be the same as in a single process environment.
A ghost values synchronization utility needs to be created
so on can propagate values among ranks given
computed ghosts. The part of this filter that needs
to be addressed for this purpose is the part using super cursors.

Fixes #19629