Disabled extract block on single htg
`vtkHyperTreeGrid` used to inherit from `vtkDataSet`. Since their
internal structure was too different, it was decided to
make it inherit from `vtkDataObject`. There was a lot of legacy
mechanism that allowed to use htg as datasets while not fully
being a dataset. `vtkSMDataTypeDomain` had a special treatment
for htgs. It is not anymore needed, hence this commit.

Also, filters taking multiblock data set input used to be hard to
handled for htgs. There was a little hack that allowed you to
consider any htg as multiblock data set, even if filters
would discard such input. This is no more needed thanks to
the distinction data set / htg.

Fixes #19862