Remove ambiguity of line seed source proxy label
On the Python side, having two proxies with the label "Line Source"
within the "extended_sources" proxy group was problematic as the
second one that was defined would overwrite the original. Fixed this
by changing the label of the line source proxy from "Line Source" to
"Line".

Follow on to #16960