Fix _camera_wrapper
The previous mechanism was creating and returning a new vtkCamera
rather than the view's active vtkCamera. Fixed that by changing the
superclass of _camera_wrapper to object.

The above change caused the, try/except mechanism for retrieving the
attribute to always exercise the except block aside from certain
methods like __dir__. Replaced this with just the except block of
code.

Override the __dir__ method to report the methods available in the
wrapped vtkCamera object as well since accessing this function
bypasses the attribute retrieval in __getattribute__().