vtkPVGeometryFilter: block-colors are paritioned-dataset-collections
When showing a vtkPartitionedDataSetCollection from an Ioss reader
in a render view, the "vtkBlockColors" array was always flagged as partial.
This was happening because the array was missing on the root-node's
field data. Since the root node has non-empty field data in case of an Ioss
reader's output, ParaView was treating the vtkBlockColors as partial.
Fixed by adding a vtkBlockColors to the root node. While it won't be
used for any rendering, it helps avoid this partial array flagging.