vtkSMOutputPort: forward calls to source-proxy
vtkSMOutputPort now forwards GetSession and GetSessionProxyManager calls
to the internal source proxy that it belongs to. This makes sense since
the output port is indeed on that session/proxy manager. This also makes
it easier for code to not have to access the source proxy just to get to
the session or proxy manager.