test-driver changes
`smTestDriver` can be passed --client-mpi instead of --client to execute
the client as if were the server with MPI, if available.

Adding `paraview_add_test` and `paraview_add_test_mpi` macros to
ParaViewTesting.cmake to make it easier to add Catalyst tests.

vtkInitializationHelper now prints out a "Process started" message when
it realizes that its being launched by a smTestDriver.