PBR: Add anisotropic and clear coat parameters
When using the PBR representation under the lighting panel, you can now
add a second layer on top of the base layer. This is useful to simulate
a coating on top of a material. This layer is dielectric (as opposed with
metallic), and can be configured with various parameters. You can set the
coat strength to control the presence of the coat layer (1.0 means
strongest coating), the coat roughness and the coat color. You can
also choose the index of refraction of the coat layer as well as the
base layer. The more the index of refraction goes up, the more specular
reflections there are.
You can also use a texture to normal map the coating.

When using the PBR representation under the lighting panel, you can now
set the anisotropy strength and anisotropy rotation for a material.
The anisotropy strength controls the amount of light reflected along the
anisotropy direction (ie. the tangent). The anisotropy rotation rotate
the tangent around the normal. Notice that the object must have normals
and tangents defined to work. You can also use a texture to hold the
anisotropy strength in the red channel, and the anisotropy rotation in
the green channel.