pqActiveObjects: support for active proxy
In preparation for adding support for a new type of pqProxy subclass to
represent extract generators that are not pqPipelineSource or have
pqOutputPort, pqActiveObjects now uses a pqProxy instead of a
pqPipelineSource or pqOutputPort to track the active pipeline object.

We also make pqOutputPort a pqProxy. vtkSMOutputPort is a vtkSMProxy
subclass, so making pqOutputPort a pqProxy subclass makes it easier to
treat it similarly.