Preview mode and text scaling.
To address paraview/paraview#20024 for the release branch, we use a
custom device pixel ratio when in preview mode. This ensures
that VTK always renders at the preview resolution despite the fact that
the Qt widget is at a different resolution. This makes it possible to
show 2D text (and other annotations) faithfully in preview mode.