Enable copying of vector and TCoords attributes
This fixes the problem where the active vectors in an input dataset
were not copied to the output while other vector arrays would be,
making it impossible to color by the orientation array if it happened
to be the active vector array in the input. Now all arrays except for
the Normals array are copied to the output.