Check IndeX library version against build version, improve logging.
The plugin will now abort startup when trying to use an IndeX library from a
different release branch than the branch against which the plugin was build.

Since an error logged during ParaView startup might not open the "Output
Messages" window, an error is now also triggered when Render() is called while
the plugin has not been initialized. This means the user will see the window
with the startup messages when switching to the IndeX representation.