Handle boundary data when running on multiple hosts (without VTK ghosting).
When running on multiple ranks, but on a single host, boundary data from a
neighboring volume piece can simply be accessed through shared memory. This
commit implements support also for multiple hosts, by transferring the requested
boundary data (not the full piece) via MPI.

Fetch requests are sent to the original owner of the neighboring piece. The
recipients of the data are always IndeX ranks, since this is where the importers
run. That means shared memory is not used here.

Not yet enabled by default. Support for double data still missing (should be
converted to float before sending).