Prevent deadlock when rendering data from .vtk file on multiple ranks.
The reader for the legacy VTK file format (.vtk) in ParaView does not support
parallel processing, it will just load the entire volume on each rank. The IndeX
plugin however still assumes Render() will get called on every rank and will
then deadlock on MPI commands when it only gets called on one rank.

This commit detects this case by checking if the extent of the current volume
piece is the same as the extent of the entire volume. If this is the case then a
warning is printed and any MPI calls skipped for rendering. The end result is
still that data will only be rendered on a single rank (=GPU).

Unstructured grid do not seem to be affected by the problem with .vtk files.