make streaming control a little more palatable
Now you can change a flag in settings/preferences and it will
take effect at runtime, with the caveat that you need to
make a new view for it to take effect. The setting now
persists across sessions so it is harder to forget.

The older command line argument is now deprecated.