pqAnimationTimeWidget: cleanup implementation
`pqAnimationTimeWidget` caused circular updates: changes in the
timesteps caused combo-box to change which had the potential to cause
the time-value to change. This manifested itself in #19953.

This fixes that by cleaning up the implementation of
pqAnimationTimeWidget to cleanly avoid causing any modifications to
ParaView state when updating UI elements.

These changes also better handle the case where the chosen timevalue is
not one of the timesteps available. The combobox now lists it as a new
entry with '(?)' suffix, a pattern we've followed elsewhere too.
The timestep spin box also uses the same suffix.