Ensure UI preserves view DPI
When the Qt-widget was created, it ended up overridding the DPI set on
the ParaView view proxy. Thus, the ParaView GUI was starting off with
DPI for 72, despite the settings/properties choosing 96. However, when
the pvbatch/pvpython respected the DPI from the properties and ended up
using 96. This was causing BUG #20217.

Fixed by ensuring that pqQVTKWidget preserves current DPI from the
window during initialization.

Since the UI was using 72 DPI, changing the default to be 72 to avoid
noticable difference in text sizes between various versions of ParaView.
That's also the default used by VTK so that keeps everything consistent.

Fixes #20217.