Add hints affecting pqDoubleRangeSliderPropertyWidget
If a hint element named "HideResetButton" is present, the range reset
button will be hidden. If a hint element named "MinimumLabel" is
present with a "text" attribute, that text attribute will be used as
the label text instead of "Minimum". Similarly, the default "Maximum"
label can be replaced with a "MaximumLabel" element with a "text"
attribute.