add an usage logger.
Adding a new behavior (pqUsageLoggingBehavior) to ping a web-service to
log each time a ParaView-based application is started. All metadata
about which server to ping and parameters to pass to it are provided
through a configuration file.

This behavior is disabled by default so that custom ParaView-based
applications don't inherit it. The behavior is explicitly enabled in
ParaViewMainWindow.

Fixes #20267.