Extractor enabled-state and eye-balls
The Pipeline Browser widget now uses the eye-balls to reflect the
extractor enabled state and change extractor enabled state. Furthermore,
for image-extractors, if the image-extractor is not associated with the
active view, the eyeball is not shown at all -- making it a little
easier to determine which image extractor is associated with which view.