Decouple GPUs used by NVIDIA IndeX from MPI ranks.
Until now the GPUs were explicitly assigned based on the MPI setup, limiting the
number of GPUs that NVIDIA IndeX could utilize to the number of MPI ranks. With
this commit, NVIDIA IndeX handles the GPU assignment internally, meaning it will
now always use all available GPUs, even if running on fewer ranks or just a
single rank.

That means running on multiple ranks on a single host is no longer necessary to
get optimum performance from NVIDIA IndeX. However, doing so might still be
beneficial to enable parallel data loading and processing in ParaView as well as
for accelerating OpenGL rendering.

Note that even before this change the "-display" parameter to pvserver did not
have an effect on the plugin, just on OpenGL rendering by ParaView.