Fix creation of new directories on server side when saving data
DirectoryProxies were being invoked on all processes when a remote
server was connected. This led to the directory being created on the
client side or reporting that the directory couldn't be created if the
directory path on the remote side didn't exist on the client side. If
the directory path did exist on the client, then the new directory
would be created there.

Fix this by setting the DirectoryProxy location to be on the servers
so that directory operations are run there only when a remote server
is connected.