Fix logic for adding layout tabs when in ReadOnly mode
In ReadOnly mode, the button to add a new tab is not present. Hence,
the tab count is too low by one. When ReadOnly mode is on, add one to
the tab count to make up for the missing tab so that logic later on
works as expected.

This resolves a problem where the first tab always stays to the right
when new layout tabs are inserted when ReadOnly mode is on.