Fix column headers for multi-component arrays
Column headers for multi-component arrays in case of composite-datasets
were not getting combined together. This was because the
"...__vtkValidMask" arrays were getting interspersed with the component
arrays causing the logic in `pqMultiColumnHeaderView` to not merge
consecutive headers. Fixed by making __vtkValidMask be a prefix,
instead of suffix. This ensures that even if the columns get sorted,
these valid mask arrays don't end up taking space in between normal
arrays and hence avoid the bug.

Fixes #20371.