Add time value as a Catalyst trigger
For simulation codes that have variable time step lengths using
the time step index as a trigger to determine when to output
Catalyst products can be a poor choice. Add in an option to
add in a time value as a Catalyst trigger which uses a
time step length as the trigger to output, i.e. try to
output Catalyst products every <X> length in simulation time units.