Improve field data handling in spreadsheet view
Spreadsheet view poorly handled showing of field data on non-leaf
nodes. This was causing #20421. To fix this, we revamp the pipeline
vtkSpreadSheetRepresentation/vtkSpreadSheetView use. Now, instead of
having to deal with a multiblock dataset in this pipeline,
vtkSpreadSheetRepresentation converts the data to a
vtkParitionedDataSet. This is done by a new data conditioning filter
called vtkDataTabulator (which replaces vtkBlockDeliveryPreprocessor).

All subsequent stages in the representation pipeline i.e.
vtkMarkSelectedRows, vtkSortedTableStreamer also now assume data
generated from vtkDataTabulator and hence get a little simplified as
well.

vtkBlockDeliveryPreprocessor is still left unchanged as its used in
vtkChartRepresentation. It should be removed in future (see #20427).

Fixes #20421.