Ensure all fields are requested in ts=0
When using V2 Catalyst Python scripts, since the script is not imported
when the first `RequestDataDescription` is called, we need to ensure all
fields/meshes are requested. The comment already mentioned that, we were
just doing toggling all the necessary flags for the fields to be
requested correctly. Fixed that.