pqDefaultContextMenu: remove block based settings
For now removing logic to show/hide blocks and set color/opacity for
blocks using the context menu. It's unclear to me with multiple
assemblies, to be supported in future, how this UI component can
continue to work robustly. It may make more sense for users to simply
reply on mutliblock-inspector.