Extractors: add support for saving recolorable images.
Adding a new extractor to save recolorable images. This still leverages
old code for capturing float values from rendering so comes with all the
idiosyncrasies that had. This just exposes that capability via the
extractor mecahnism. The extractor is added to "Experimental" category
to make clear it is still not ready for production use and only
available for advanced users.