cmake: offer a variable to indicate that ParaView will be built
In !4755, testing was taught to check whether the `ParaView::paraview`
target exists or not rather than relying on `PARAVIEW_USE_QT`. While
this is a correct change in terms of asking what is actually needed,
within ParaView's build, it is not exactly correct because ParaView
itself might not have been scheduled for build when it is asked for
(primarily within plugins and module builds). Instead, offer a local
variable which indicates that ParaView will be built and use it as well.

It is kept with the target check in order to support testing given an
external ParaView in the future.