Handle vtkGhostType arrays specially in vtkSIDataArraySelectionProperty
Ghost type arrays are kind of oddballs - some filters such as
vtkPassSelectedArrays assume that if a ghost array is not explicitly
marked disabled that it should be treated as enabled. Make sure we
initially explicitly disable these in a selection, then re-enable them
if they are listed in the array selection property.