vtkSMProperty: avoid updating domains on prototypes
vtkSMProperty no longer updates dependent domains when the property is
modified on prototype proxies. This is largely unnecessary given how
prorotype proxies are used in ParaView.

This fixes #20617. The warning was showing up since the pipeline browser
was trying to see if Ioss reader is an acceptable input for
chart-representation. That check was causing chart representation
prototype to attempt to update its other property domains raising the
warnings.