Support NVIDIA IndeX plugin on macOS (remote rendering only).
When a ParaView client connects to a remote pvserver, the NVIDIA IndeX plugin
must be loaded on both the client and the server. Since the plugin was not
available on macOS so far, it was not possible to connect from a macOS client to
a Linux server.

This commit makes it possible to build the plugin on macOS, but disables the
actual loading of IndeX libraries (which are not available on macOS). Remote
rendering, i.e. connecting to a pvserver process running on Linux, now works.
When the user tries to enable the "NVIDIA IndeX" representation locally, an
error will be printed.