Fix TimeArrayInfo property on readers
The information property TimeArrayInfo for the TimeArray property on
XML and PIO readers was a list of possible time arrays from the
dataset, not the selected TimeArray.  In certain situations, this
could lead to the TimeArray property being initialized to the list of
possible time arrays, which is not valid behavior.

Fixed by removing the TimeArrayInfo property as the information property
of the TimeArray property.