Create and register a 2D transfer function proxy with the controller
The transfer function manager creates and registers a new 2D transfer
function proxy based on the XML spec with the controller. This makes it
available to both the volume representation and the transfer function
editor. Effectively, this change makes it so that the same transfer
function being edited by the editor is used for mapping by the
representation proxies.