Cleanup background color
Instead of using mutliple properties to setup background mode with the
potential for conflicting options, we now use a single
"BackgroundColorMode" property for render views to control how the
background is selected.

This also make custom widget for the background color obsolete as we can
simply use property-widget-decorators to control how the widgets are
shown in the properties panel.