Explicit view background - color palette links
To minimize confusion when views are linked or not linked with global
color palette for background settings, we now add an explicit checkbox
that the user can toggle to indicate if background should be liked to
global palette or not. Unless unchecked, the view specific parameters
are ignored and the view simply uses global settings.

Addresses issues raised in #15757.