Update OpenVR plugin due to VTK collab changes
Collaboration support is now an option in VTK so this
plugin now turns on/off collaboration based on the VTK
setting. It no longer has any dependency on mineview
source code. VTK collab support does require libzmq so
if that option is turned on it will try to find libzmq.