Fixing legacy bug for Plot Over Line
Plot Over Line backen changed recently, and its proxy API changed.
State files (.pvsm) produced with master between 5.9 and 5.10 wrongly
downgraded the proxy to be loaded to the legacy one. This is because
current builds are tagged as 5.9, even if it already has changes that
are targetted for 5.10. Since the proxy API is different, we can
discriminate "true" 5.9 instances of .pvsm files from the ones built
with master after the backend was changed. In particular, `Point1` and
`Point2` properties are new properties that didn't use to be there.

Fixes #20778