Looking glass plugin: add ability to save quilts
This MR depends on [this LookingGlassVTKModule PR](https://github.com/Kitware/LookingGlassVTKModule/pull/20).

Two buttons are added to the looking glass plugin dock panel: one for
saving quilt images, and one for starting/stopping recording for quilt
movies.

If the "Save Quilt" button is pressed, a PNG file with the name
"quilt_qsMxN.png" (where `M` is the number of rows and `N` is the
number of columns) will be saved in the user's working directory.
The `_qsMxN` suffix communicates the dimensions of the quilt to the
LG software that loads it.

If the "Record Quilt" button is pressed, recording a quilt movie
begins, and the text of the button changes to "Stop Recording Quilt"
(where pressing it a second time stops and finalizes the recording).
During recording, each time a render is performed to the LG device,
a frame is written to the quilt movie. Thus, a user may control the
frames that get written to the movie file by controlling when each
render to the LG device happens (which the user can do via the
"Rendering" combo box in the plugin's dock panel).

The movie file is named "quilt_qsMxN.*", where `*` will be "mp4",
"avi", or "ogg", depending on which movie writing software is
available in VTK. MP4 is preferred since it is the only format
that LG software can read in, but it is only available on Windows.
Thus, on other platforms, the user must convert the AVI or OGG
file to a format LG software can read.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>