Add option to change quilt export magnification
The quilts that get exported as images/videos typically need to be
magnified in order to appear nice on an LG device.

It seems that a magnification of 2 is working well, so that is the
default. But the magnification option is now exposed so that the user
may modify it if needed.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>