catalyst: transport Catalyst API variables around
ParaView doesn't use `find_package(catalyst)` in the right scope for
everything (since it is optionally embedded), so instead shuttle the
variable to the necessary places via a global property.

Logic for external Catalyst versions is added to minimize the chances of
things going wrong if/when an external Catalyst is supported.

Fixes: #20796