Looking Glass plugin: choose filenames for quilts
This brings up save file dialogs for both the image and video quilts
that allow the user to choose both the path and the name of the quilt.

HoloPlay Studio uses information encoded into the name of the quilt
to determine the number of tiles. As such, a suffix, such as
"_qs5x9", will automatically be added to the name of the quilt
immediately before the file extension.

There is still a dialog that asks if the user wants to overwrite the
file if the file (with the suffix added) already exists.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>