Resample to htg: 2D bug fix
Dimension indexing was wrongly handled in the 2D case, causing the
filter to crash / produce a wrong output.

Crashes were happening when the user would put wrong dimensions for the
output htg, for instance setting one to 0, instead of 1. Now, if the
user sets a wrong dimension, it is reset to 1.

In order to correct the indexing bug for 2D, checking if a dimension is
degenerate when setting the coordinates (i,j,k) fixes the problem.

The test is being updated to enfore 2D outputs to work.

Fixes #20787