Change some behaviors of the Color Map Editor
* Add a combo box to quickly select a preset from the Default group directly
from the Color Map Editor.

* Change the way presets groups are working. Previously, the groups were always
loaded from the pqPresetGroups.json. This was used to create the categories used
by the presets dialog. It is now only used when its equivalent cannot be found
in the settings under the "TransferFunctionPresets.Groups" key. For the Default
group, it is now used over the previous "pqSettingdDialog/userChosenPresets" key
which did not allow the removal of presets from the Default group, which is now
possible

* Automatically add the imported presets to a User group as well as the Default
group.

* Make the presets dialog non-modal