Use standard "Cropping Parameters" in NVIDIA IndeX plugin.
ParaView 5.8 added a UI for setting "Cropping Parameters", which can also be
controlled by moving widgets on the rendering canvas. This is currently only
available for structured grids.

This commits connects the "Cropping" settings to NVIDIA IndeX and removes the
custom "Region of Interest" UI for structured grids. It is still in place for
unstructured grids.

Old state files will continue to work, but a warning will be printed, showing
the corresponding settings for Cropping Origin and Cropping Scale:
  The 'NVIDIA IndeX Region of Interest' setting is deprecated, please use
  'Cropping' instead (origin: [0 5 0], scale: [1 0.5 1]).