Fix "Update kernel" not always updating rendering in NVIDIA IndeX plugin.
While changing properties will automatically trigger rendering, pressing the
"Update kernel" button does not. This has the effect that the XAC shaders will
not be reloaded and updated, which is most apparent when XAC compilation of a
user-defined shader has failed and after fixing the bug in the shader file,
pressing "Update kernel" does not change the rendering. Only updating any other
property or moving the camera will trigger the rendering and therefore
recompilation of the shader.

This commit explicitly calls vtkPVRenderView::StillRender() when the button is
pressed and the kernel source has been changed.