pqRenderView: support BLOCK_SELECTORS selection type
When selecting blocks in a PartitionedDataSetCollection, pqRenderView
now creates a "selector" based selection. This makes it possible to
correctly select a block i.e. an entire PartitionedDataSet in a when
user selects it rather than an single partition. This also makes the
selection consisent when running in parallel since in parallel composite
ids for PDCs are inconsistent across ranks and any composite-id based
selection is incorrect.

Fixes #20904.