DataInformation improvements:
Adding mechanism to collect data information from a particular rank.
This becomes necessary as the composite hierarchy across ranks need not
be consistens (e.g. for PDCs, and PDs) and we make require data
information from a specific rank.

Also adds some API to vtkPVDataInformation to make it easier to deal
with composite datasets and hierarchies.