Bring back context menu for block visibilities etc.
With introduction of PartitionedDataSet and
PartitionedDataSetCollection, composite ids are no longer consistent
across ranks. So to support use-cases like #20899, we have to take
extra care to ensure we're referring to correct block by using rank
information as well.

This commit brings back the ability to setup block controls in pipeline
context menu. pqContextMenuInterface was extended to add rank
information.

Fixes #20899.