comments: rework comments after reformatting
Many of these comments were not wordwrapped well, so that is the main
change being made here. In addition, some comments were pushed around to
align with other comments that just happened to be on adjacent lines.
These comments are moved around so that the formatting ends up looking a
lot better.