Add web page and movie representations to OpenVR plugin
The vtkWebPageRepresentation allows one to use web pages
as representations in ParaView. It makes use of the QtWebEngine
to perform all the heavy lifting and mearly handles display
and event forwarding. The web page to display is entered as
text in a table (like the text representation)

Add movie rep which allows movies to be played as skybox
backgrounds in ParaView.