Workaround for NVIDIA IndeX plugin shutdown issue on Windows.
After the NVIDIA IndeX libraries were upgraded to 348900.100.964 in MR !5143,
the ParaView process started exiting with a non-zero exit code on Windows, when
the IndeX plugin was active and a dataset loaded. This happened because
the (undefined) order of static variable destruction had changed.

This workaround goes back to the old behavior in ParaView 5.9 (before commit
13edd14ab9c7b59f343228c8dbbb311e2e803c22), and now won't explicitly shut down
the IndeX instance, intentionally leaking the pointer. A better solution will
require a mechanism to react to ParaView shutting down, which currently does not
seem to be available.