avoid making non-updated representations as updated
`vtkSMViewProxy::Update` caused all representations (and their upstream)
pipelines to think they they were updated. This is generally true,
except if the representation was hidden. This incorrect notification for
hidden representation meant that unless the pipeline was explicitly
dirtied the data information would not be updated correctly. Fixed that.

Fixes #20133.